MBKettouExec decomp

+++ src/overlays/ovl_48_ChillyWaters/ChillyWaters.h
@@ -94,7 +94,7 @@ extern char D_8011E344_333EB4_ChillyWaters[]; //unknown type
 extern char D_8011E4D8_334048_ChillyWaters[]; //unknown type
 extern char D_8011E718_334288_ChillyWaters[]; //unknown type
 extern s16 D_8011D2C0_332E30_ChillyWaters[8];
-extern s32 D_8011D31C_332E8C_ChillyWaters; //unknown type
+extern s16 D_8011D31C_332E8C_ChillyWaters; //unknown type
 extern s32 D_8011D320_332E90_ChillyWaters; //unknown type
 extern s16 D_8011D2A0_332E10_ChillyWaters[8];
 extern s16 D_8011D2B0_332E20_ChillyWaters[8];
